Fix double task release on shutdown
Back when we made Run's return value (done bool) in 26274e5 -- that
implicitly made Run the sole arbiter of whether or not to Release a task
or Mark it as done.

The only thing the core Consumer should do is tell handlers to Stop --
it should only ever call Release or Done after Run exits and use that
return value.

We do lose the ReleaseTask command (since now it's the same as the
StopTask command), but we've never used it and it can be added later if
it's found to be valuable.