CanClaim rejects tasks for a period of time.
**API breaks: Coordinator and Balancer**

This change makes Metafora look uglier in the abstract but actually
handle operational realities better. The existing CanClaim interface
didn't reflect reality: if it rejected a task the Coordinator's Watch
method might just continue returning only that task forever...and in a
tight loop no less! There was no way for the Coordinator's Watch and
Balancer to communicate (unless they implemented shared state outside
the Consumer which sounds really tricky to get right).

This design makes Coordinator.Watch stream claimable tasks back to the
Consumer. The Consumer checks an ignore-list before even calling
Balancer.CanClaim. If a task isn't ignored, CanClaim can cause it to be
ignored until a deadline.

Once the deadline is up *the Consumer handles the task as if it came
from the Coordinator.* This means it gives the Balancer's CanClaim
method another chance to reject it and everything.

Even if `CanClaim` doesn't reject the task again, the task probably
can't be claimed anyway because another node claimed it ages ago.
That's ok. This design trades off being optimally efficient for safety:
it never rejects a task forever (possibly orphaning it), and it never
bypasses the Balancer.

Fixes #93

Errors returned by `Coordinator.Watch` or `Coordinator.Command` are now
**fatal**. Nowhere in our logs could I find an instance of them failing
(even in our flaky staging environment).  I think it's fine to expect
the Coordinator to be resilient and only return errors when *it*
determines it cannot continue to function safely.

Fixes #71

Also makes `DumbBalancer` a singleton because it's silly not to.