Sleep after successful claims to encourage fairness
It was observed when restarting a cluster that if a node won the first
claim it would win *all* subsequent claims as claiming and failing to
claim take nearly identical resources.

This slows down the "winner" to try to encourage a fairly balanced
cluster.

While balancers are probably a better place to do this, the consumer
should prevent pathological behaviors like we were seeing, so I think a
slight sleep in the core consumer is fine.