Fix race when adding ignore during shutdown
Previously if Consumer.Shutdown was called while the Consumer was adding
an ignored task, the ignore manager's send of the task to the
ignored-task-monitor would block forever -- preventing shutdown of the
consumer.

This change allows the sending of the ignored task to the monitor to
abort if a shutdown has been initiated. Note that this leaves the
ignoremgr's map and timeheap in an inconcistent state during shutdowns,
but there's no case in which they're used during shutdown so it's safe.