Fix bad behavior where fair balancer would spin after releasing tasks
The problem:

* `FairBalancer.Balance()` releases tasks and sets the internal delay
* Coordinator tries to claim rebalanced tasks, Claim fails, sets ignore
  * This step repeats until `Balance()` is called again

The affects:

* Spams logs (because often only 1 or 2 tasks are released)
* Prevents any new claims after a rebalance which could cause a
  misbalanced cluster and more balancing!

The fix:

Set an *absolute* time for the delay instead of a relative time. Once
the delay expires, all tasks can be claimed even if they're the ones
recently released.

This should prevent a node from reclaiming tasks it *just* released, but
it doesn't totally prevent the node from claiming tasks until the next
Balance call.