Make sure a successful claim didn't resurrect a deleted task
This introduces an extra GET (and potentially DELETE) into the task
claim flow which is unfortunate. I don't see another way of guaranteeing
a claim can't recreate a deleted task without the extra check though.

Looks like we're going to have to wait for a new etcd in order to get a
more efficient way of doing this:

https://github.com/coreos/etcd/issues/2857

Fixes #124