Make stateful handler command chan very defensive
I don't like this code, but that doesn't mean it's not a good idea. It's
just very easy for a handler to accidently orphan a goroutine that
receives commands. That effectively steals commands from e.g. sleep's
select cases.

A better solution might be a totally different API for sending commands
to tasks (perhaps a Command(m Message) method for handlers to implement.