implement error interface for statemachine.Err (#174)
This commit implements the error interface for statemachine.Err. The change preserves the existing external interface, implements the error interface, and adds a private field baseErr so that the type is a proper error and can be converted back to its underlying error type using errors.As and errors.Is. A constructor is added to simplify instantiation and to allow for testing outside of the statemachine package, since we don't want to expose the underlying error directly.

Also updates the go directive in go.mod to 1.13, and requires 1.13.x in .travis.yml since we're now using features in the errors package that were introduced in 1.13

* embed error in statemachine.Err and implenent custom As() so that we preserve the underlying error without changing the existing external interface

* update metafora to go 1.13: this is necessary if we want to take advantage of the new error patterns

* use go 1.13 for travis builds

* don't embed: bad idea, Error() will panic if someone instantiates statemachine.Err the old-fashioned way (without an underlying error) and decides to call Error() for some reason. Would be an odd choice, but seems plausible as an err != nil check will return true in that case

* improve comments to provide context/history behind the change

* Update statemachine/errors.go

Co-Authored-By: Matt Braymer-Hayes <matthew.braymer-hayes@lytics.com>

* Update statemachine/errors_test.go

Co-Authored-By: Matt Braymer-Hayes <matthew.braymer-hayes@lytics.com>

* Update statemachine/errors_test.go

Co-Authored-By: Matt Braymer-Hayes <matthew.braymer-hayes@lytics.com>

* Update statemachine/errors.go

Co-Authored-By: Matt Braymer-Hayes <matthew.braymer-hayes@lytics.com>

* implement Unwrap instead of As, add test cases to ensure errors.Is works as expected

Co-authored-by: Matt Braymer-Hayes <matt.hayes91@gmail.com>